Post and Get functions

+++ Backend/Controllers/ValuesController.cs
@@ -81,7 +81,7 @@ namespace ProductsAPI.Controllers
             public int? Id { get; set; }
             public string? Name { get; set; }
             public string? Category { get; set; }
-            public int? Price { get; private set; }
+            public float? Price { get; private set; }
 
             public void UptatePrice(int? NewPrice)
             {
